Move form-utils from @console/dev-console to @console/shared
