+++ frontend/packages/dev-console/src/components/pipelines/detail-page-tabs/PipelineParameters.tsx
@@ -5,7 +5,7 @@ import { FormikProps, FormikValues } from 'formik';
 import { useAccessReview } from '@console/internal/components/utils';
 import { getActiveNamespace } from '@console/internal/actions/ui';
 import { MultiColumnField, InputField } from '@console/shared/src/components/formik-fields';
-import { FormFooter } from '../../form-utils';
+import { FormFooter } from '@console/shared/src/components/form-utils';
 
 const PipelineParameters: React.FC<FormikProps<FormikValues>> = ({
   handleSubmit,
