+++ frontend/packages/dev-console/src/components/pipelines/detail-page-tabs/PipelineResources.tsx
@@ -9,7 +9,7 @@ import {
   InputField,
   DropdownField,
 } from '@console/shared/src/components/formik-fields';
-import { FormFooter } from '../../form-utils';
+import { FormFooter } from '@console/shared/src/components/form-utils';
 
 enum resourceTypes {
   '' = 'Select resource type',
